minor bug fixes

+++ bot.py
@@ -29,6 +29,9 @@ QUEUE=queue.Queue()
 #prefix to signal the bot to listen to the command
 client = commands.Bot(command_prefix= '!', )
 
+#state variable in rare cases where the song is loaded and you hit the idle timer
+loading_song_state=False
+
 ### Starting message to give some feedback that bot has started
 @client.event
 async def on_ready():
@@ -41,20 +44,19 @@ async def on_ready():
 @client.command(pass_context = True)
 async def play(ctx, *, arg):
     voice = discord.utils.get(client.voice_clients, guild=ctx.guild)
-    if (voice and ctx.message.author.voice.channel != voice):
-        print("can't play songs while not in the channel")
+    if (voice and ctx.message.author.voice.channel != voice.channel):
+        print("can't play songs while user is not in same channel as bot")
         return
-    if voice and not leaveLoop.is_running():
-        leaveLoop.start(voice, ctx)
     url = getYTURL(arg)
     #if a song is playing or queue is already filled up, add to the back of queue
     if voice != None and (QUEUE.qsize() >= 1 or voice.is_playing()):
         await ctx.send("added: " + url + " to the queue")
+    print("adding to queue")
     QUEUE.put(url)
 
     if(voice == None or not voice.is_playing()):
+        print("inside if no voice")
         try:
-            url = getYTURL(arg)
             #check if bot is in current channel, move it if needed
             if(ctx.author.voice):
                 channel = ctx.message.author.voice.channel
@@ -62,10 +64,7 @@ async def play(ctx, *, arg):
                     voice = await channel.connect()
                     if(not leaveLoop.is_running()):
                         leaveLoop.start(voice, ctx)
-            source = downloadAndGetSource(voice)
-
-            #finally play the song and recursivly play the next songs in the queue
-            voice.play(FFmpegPCMAudio(source=source), after=lambda e: playNext(ctx))
+            await playNext(ctx)
 
             #give user feedback with embedded link to the video playing
             await ctx.send("fine I'll play " + url + "... you're welcome :rolling_eyes:")
@@ -79,9 +78,9 @@ async def skip(ctx):
     if(QUEUE.qsize() == 0 and not voice.is_playing()):
         await ctx.send("no song playing to skip and no songs in queue...")
     else:
-        await voice.stop()
+        voice.stop()
         await ctx.send("yeah, that song was :peach:... Skipping...")
-        playNext(ctx)
+        await playNext(ctx)
 
 ### really a test function to force the bot into the current channel
 @client.command(pass_context = True)
@@ -141,6 +140,8 @@ def checkFileExists(voice):
 
 ### downloads the song from YT and stores it to be played
 def downloadAndGetSource(voice):
+    global loading_song_state
+    loading_song_state = True
     #check if the bot is currently playing stop it and reset for next song if it is
     checkFileExists(voice)
     with youtube_dl.YoutubeDL(YDL_PROPS) as ydl:
@@ -149,6 +150,7 @@ def downloadAndGetSource(voice):
         if file.endswith(FILE_EXTENSION):
             os.rename(file, SONG_FILE)
     source = SONG_FILE
+    loading_song_state = False
     return source
 
 ### forms the youtube query url
@@ -160,20 +162,23 @@ def getYTURL(arg):
     return YT_WATCH_BASE_URL + video_ids[0]
 
 ### recursive call to continue to play the queue
-def playNext(ctx):
+async def playNext(ctx):
+    print("in next")
     voice = discord.utils.get(client.voice_clients, guild=ctx.guild)
     if QUEUE.qsize() >= 1:
         source = downloadAndGetSource(voice)
         voice.play(FFmpegPCMAudio(source=source), after=lambda e: playNext(ctx))
 
 ### once a song starts check back in every 3 minutes to see if bot needs to disconnect
-@tasks.loop(minutes=3)
+@tasks.loop(minutes=1)#minutes=3)
 async def leaveLoop(voice, ctx):
-    if not voice.is_playing() and QUEUE.qsize() == 0:
+    if not voice.is_playing() and not voice.is_paused() and QUEUE.qsize() == 0 and not loading_song_state:
         if(ctx.voice_client):
-            await ctx.guild.voice_client.disconnect()
-            os.remove(PATH_TO_SONG_FILE)
-            await ctx.send("fuck it I'm out :peace:")
+            if(voice.is_connected()):
+                await ctx.send("fuck it I'm out :peace:")
+                await ctx.guild.voice_client.disconnect()
+            if(exists(PATH_TO_SONG_FILE)):
+                os.remove(PATH_TO_SONG_FILE)
             leaveLoop.stop()
 
 client.run(APIKEY)
